Int attrib

+++ OpenGL/res/shaders/default.vert
@@ -3,11 +3,11 @@
 layout(location = 0) in vec4 a_Position;  
 layout(location = 1) in vec4 a_Color;
 layout(location = 2) in vec2 a_TexCoord;
-layout(location = 3) in float a_TexID;
+layout(location = 3) in int a_TexID;
 
 out vec4 v_Color;
 out vec2 v_TexCoord; // v = varying
-out float v_TexID;
+out int v_TexID;
 
 uniform mat4 u_ViewProj;
